refactor(backend): DTO 파일명을 모듈 네이밍 컨벤션에 맞게 수정
